+++ src/labkey/Query.ts
@@ -37,21 +37,25 @@ import {
     deleteRows,
     insertRows,
     saveRows,
-    selectDistinctRows,
-    selectRows,
     truncateTable,
     updateRows
 } from './query/Rows'
+import { selectDistinctRows } from './query/SelectDistinctRows'
+import { selectRows } from './query/SelectRows'
 import * as SQL from './query/experimental/SQL'
 import * as Visualization from './query/Visualization'
 import { Filter } from './filter/Filter'
-import { Response, Row } from './query/Response';
+import { Response, Row } from './query/Response'
+import { QueryColumn } from './query/types'
 
 const experimental = {
     SQL
 };
 
 export {
+    // Interfaces
+    QueryColumn,
+
     ContainerFilter, // Enumeration
     containerFilter, // backwards compatible reference
     buildQueryParams,